+++ pkg/network/protocols/kafka/model_linux.go
@@ -23,12 +23,12 @@ func (tx *EbpfTx) ConnTuple() types.ConnectionKey {
 
 // APIKey returns the API key for the transaction
 func (tx *EbpfTx) APIKey() uint16 {
-	return tx.Request_api_key
+	return uint16(tx.Request_api_key)
 }
 
 // APIVersion returns the API version for the transaction
 func (tx *EbpfTx) APIVersion() uint16 {
-	return tx.Request_api_version
+	return uint16(tx.Request_api_version)
 }
 
 // RecordsCount returns the number of records in the transaction
